+++ src/components/OverlayWrapper/tests/index.js
@@ -1,6 +1,9 @@
 import React from 'react';
-import OverylayTrigger from '../index';
+import OverylayWrapper from '../index';
 import { shallow, mount } from 'enzyme';
+
+jest.mock('tether');
+jest.useFakeTimers();
 
 const FakeButton = (props) => {
   return <button data-test-section="fake-button"></button>;
@@ -13,30 +16,30 @@ const FakeOverlay = (props) => {
 describe('components/OverlayWrapper', () => {
   it('should render contents passed in as children', () => {
     const component = mount(
-      <OverylayTrigger
+      <OverylayWrapper
         overlay={ <FakeOverlay /> }>
         <FakeButton />
-      </OverylayTrigger>
+      </OverylayWrapper>
     );
     expect(component.find('[data-test-section="fake-button"]').length).toBe(1);
   });
 
   it('should render overlay', () => {
     const component = mount(
-      <OverylayTrigger
+      <OverylayWrapper
         overlay={ <FakeOverlay /> }>
         <FakeButton />
-      </OverylayTrigger>
+      </OverylayWrapper>
     );
     expect(component.find('[data-test-section="fake-overlay"]').length).toBe(1);
   });
 
   it('should show or hide the `overlay` depending on state', () => {
     const component = mount(
-      <OverylayTrigger
+      <OverylayWrapper
         overlay={ <FakeOverlay /> }>
         <FakeButton />
-      </OverylayTrigger>
+      </OverylayWrapper>
     );
 
     // Start off with the default state
@@ -51,65 +54,135 @@ describe('components/OverlayWrapper', () => {
     expect(component.instance()._overlayEl.getAttribute('style')).toContain('display: block');
   });
 
-  it('should pass `onClick` function to `children` that changes state of ``', () => {
-    spyOn(React, 'cloneElement');
+  describe('clicking on the children', () => {
+    let fakeButtonClone;
+    let component;
+    const foo = { bar: () => {} };
 
-    const component = shallow(
-      <OverylayTrigger
-        overlay={ <FakeOverlay /> }>
-        <FakeButton />
-      </OverylayTrigger>
-    );
+    beforeEach(() => {
+      spyOn(React, 'cloneElement').and.callThrough();
+      spyOn(foo, 'bar');
 
-    const fakeButtonClone = React.cloneElement.calls.mostRecent();
+      component = mount(
+        <OverylayWrapper
+          overlay={ <FakeOverlay /> }>
+          <FakeButton onClick={ foo.bar } />
+        </OverylayWrapper>
+      );
 
-    // Ensure that the default state is what this test expects
-    expect(component.state('isOverlayOpen')).toBe(false);
-    // Ensure that we are looking at the correct call to `React.cloneElement`
-    expect(fakeButtonClone.args[0].type).toBe(FakeButton);
+      fakeButtonClone = React.cloneElement.calls.mostRecent();
+    });
 
-    // Simulate a click on the button
-    fakeButtonClone.args[1].onClick();
+    afterEach(() => {
+      React.cloneElement.calls.reset();
+      foo.bar.calls.reset();
+      component.unmount();
+      fakeButtonClone = null;
+      component = null;
+    });
 
-    // Expect that the state changed
-    expect(component.state('isOverlayOpen')).toBe(true);
+    it('should still call `children`\'s existing `onClick` if it exists', () => {
+      // Ensure that we are looking at the correct call to `React.cloneElement`
+      expect(fakeButtonClone.args[0].type).toBe(FakeButton);
+
+      // Simulate a click on the button
+      fakeButtonClone.args[1].onClick();
+
+      // Expect that `FakeButton`'s `onClick` was called
+      expect(foo.bar.calls.count()).toBe(1);
+    });
+
+    it('should enable and disable Tether when clicking on the children', () => {
+      // Ensure that we are looking at the correct call to `React.cloneElement`
+      expect(fakeButtonClone.args[0].type).toBe(FakeButton);
+
+      // Store the mock for easy access
+      const disableMock = component.instance()._tether.disable.mock;
+      const enableMock = component.instance()._tether.enable.mock;
+      const initialDisableMockCalls = disableMock.calls.length;
+      const initialEnableMockCalls = enableMock.calls.length;
+
+      // Simulate a click on the button
+      fakeButtonClone.args[1].onClick();
+
+      // Tether should now be enabled
+      expect(enableMock.calls.length).toBe(initialEnableMockCalls + 1);
+
+      // Simulate a click on the button
+      fakeButtonClone.args[1].onClick();
+
+      // Tether should now be disabled
+      expect(disableMock.calls.length).toBe(initialDisableMockCalls + 1);
+    });
+
+    it('should position Tether when clicking on the children', () => {
+      // Ensure that we are looking at the correct call to `React.cloneElement`
+      expect(fakeButtonClone.args[0].type).toBe(FakeButton);
+
+      // Store the mock for easy access
+      const positionMock = component.instance()._tether.position.mock;
+      const initialPositionMockCalls = positionMock.calls.length;
+
+      // Simulate a click on the button
+      fakeButtonClone.args[1].onClick();
+
+      // Run the `setTimeout` synchronously
+      jest.runAllTimers();
+
+      // The `Tether.position` function should have been called once
+      expect(positionMock.calls.length).toBe(initialPositionMockCalls + 1);
+    });
+
+    it('should pass `onClick` function to `children` that changes state of `OverylayWrapper`', () => {
+      // Ensure that we are looking at the correct call to `React.cloneElement`
+      expect(fakeButtonClone.args[0].type).toBe(FakeButton);
+
+      // Ensure that the default state is what this test expects
+      expect(component.state('isOverlayOpen')).toBe(false);
+      // Ensure that we are looking at the correct call to `React.cloneElement`
+      expect(fakeButtonClone.args[0].type).toBe(FakeButton);
+
+      // Simulate a click on the button
+      fakeButtonClone.args[1].onClick();
+
+      // Expect that the state changed
+      expect(component.state('isOverlayOpen')).toBe(true);
+    });
   });
 
-  it('should still call `children`\'s existing `onClick` if it exists', () => {
-    let foo = { bar: () => {} };
-
-    spyOn(React, 'cloneElement');
-    spyOn(foo, 'bar');
-
-    shallow(
-      <OverylayTrigger
+  it('should disable Tether by default', () => {
+    const component = mount(
+      <OverylayWrapper
         overlay={ <FakeOverlay /> }>
-        <FakeButton onClick={ foo.bar } />
-      </OverylayTrigger>
+        <FakeButton />
+      </OverylayWrapper>
     );
 
-    const fakeButtonClone = React.cloneElement.calls.mostRecent();
+    const disableMock = component.instance()._tether.disable.mock;
+    const enableMock = component.instance()._tether.enable.mock;
 
-    // Ensure that we are looking at the correct call to `React.cloneElement`
-    expect(fakeButtonClone.args[0].type).toBe(FakeButton);
-
-    // Simulate a click on the button
-    fakeButtonClone.args[1].onClick();
-
-    // Expect that `FakeButton`'s `onClick` was called
-    expect(foo.bar.calls.count()).toBe(1);
+    expect(disableMock.calls.length).toBe(1);
+    expect(enableMock.calls.length).toBe(0);
   });
 
   describe('passing options to Tether', () => {
+    beforeEach(() => {
+      spyOn(OverylayWrapper.prototype, 'createTether').and.callThrough();
+    });
+
+    afterEach(() => {
+      OverylayWrapper.prototype.createTether.calls.reset();
+    });
+
     it('should pass the correct options with none of the layout props provided', () => {
       const component = mount(
-        <OverylayTrigger
+        <OverylayWrapper
           overlay={ <FakeOverlay /> }>
           <FakeButton />
-        </OverylayTrigger>
+        </OverylayWrapper>
       );
 
-      const tetherOptions = component.instance()._tether.options;
+      const tetherOptions = OverylayWrapper.prototype.createTether.calls.mostRecent().args[0];
 
       expect(tetherOptions.attachment).toBe('top center');
       expect(tetherOptions.constraints.length).toBe(1);
@@ -118,16 +191,13 @@ describe('components/OverlayWrapper', () => {
         to: 'window',
         pin: false,
       });
-      expect(tetherOptions.offset).toBe('0 0');
-      expect(tetherOptions.targetAttachment).toBe('auto auto');
-      expect(tetherOptions.targetOffset).toBe('0 0');
       expect(tetherOptions.target.tagName.toLowerCase()).toBe(component.find(FakeButton).render().children().first()[0].name);
       expect(tetherOptions.element.tagName.toLowerCase()).toBe(component.find(FakeOverlay).render().children().first()[0].name);
     });
 
     it('should pass the correct options with all the layout props provided', () => {
       const component = mount(
-        <OverylayTrigger
+        <OverylayWrapper
           overlay={ <FakeOverlay /> }
           horizontalAttachment="center"
           verticalAttachment="top"
@@ -135,10 +205,10 @@ describe('components/OverlayWrapper', () => {
           horizontalTargetAttachment="center"
           isConstrainedToScreen={ true }>
           <FakeButton />
-        </OverylayTrigger>
+        </OverylayWrapper>
       );
 
-      const tetherOptions = component.instance()._tether.options;
+      const tetherOptions = OverylayWrapper.prototype.createTether.calls.mostRecent().args[0];
 
       expect(tetherOptions.attachment).toBe('top center');
       expect(tetherOptions.constraints.length).toBe(1);
@@ -147,9 +217,7 @@ describe('components/OverlayWrapper', () => {
         to: 'window',
         pin: true,
       });
-      expect(tetherOptions.offset).toBe('0 0');
       expect(tetherOptions.targetAttachment).toBe('top center');
-      expect(tetherOptions.targetOffset).toBe('0 0');
       expect(tetherOptions.target.tagName.toLowerCase()).toBe(component.find(FakeButton).render().children().first()[0].name);
       expect(tetherOptions.element.tagName.toLowerCase()).toBe(component.find(FakeOverlay).render().children().first()[0].name);
     });
@@ -158,10 +226,10 @@ describe('components/OverlayWrapper', () => {
 
   it('should destroy Tether on unmount', () => {
     const component = mount(
-      <OverylayTrigger
+      <OverylayWrapper
         overlay={ <FakeOverlay /> }>
         <FakeButton />
-      </OverylayTrigger>
+      </OverylayWrapper>
     );
 
     const instance = component.instance();
@@ -176,11 +244,11 @@ describe('components/OverlayWrapper', () => {
 
   it('should have a properly set test section', () => {
     const component = shallow(
-      <OverylayTrigger
+      <OverylayWrapper
         testSection="foo"
         overlay={ <FakeOverlay /> }>
         <FakeButton />
-      </OverylayTrigger>
+      </OverylayWrapper>
     );
     expect(component.is('[data-test-section="foo"]')).toBe(true);
   });